projects/dynamic-form/src/lib/components/input/input.component.ts
import { Component, Input } from '@angular/core';
import { FormControl } from '@angular/forms';
import { InputConfiguration } from '../../models/models';
import { errorControlMessage } from '../../utils/errors/error-control';
import { hasRequiredValidator } from '../../utils/validators/has-validator';

@Component({
  selector: 'talan-input',
  templateUrl: './input.component.html',
  styleUrls: ['./input.component.scss'],
})
export class InputComponent {
  private _field!: InputConfiguration;

  get field(): InputConfiguration {
    return this._field;
  }

  @Input() set field(value: InputConfiguration) {
    this._field = value;
  }

  @Input() control!: FormControl;

  get required() {
    return hasRequiredValidator(this.control);
  }

  constructor() {}

  errorControlMessage = errorControlMessage;
}
